Move non-shared out of shared
